projects/file_analyzer/requirements.txt
pandas==2.0.3
matplotlib==3.7.2
python-magic-bin==0.4.14
